Sample Controller クラス移動
