+++ src/Controller/Other/SampleCode/MySqlTypeSamples/EditController.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Controller\SampleCode\MySqlTypeSamples;
+namespace App\Controller\Other\SampleCode\MySqlTypeSamples;
 
 use App\Controller\AppController;
 use App\Exception\ValidateException;
 use App\Security\Auth\AuthContextResolver;
-use App\Service\Controller\SampleCode\MySqlTypeSamples\Edit as CtlService;
+use App\Service\Controller\Other\SampleCode\MySqlTypeSamples\Edit as CtlService;
 use Cake\Event\EventInterface;
 use Cake\Http\Response;
 use DateTimeImmutable;
@@ -14,7 +14,7 @@ use DateTimeImmutable;
 class EditController extends AppController
 {
     /**
-     * @var \App\Service\Controller\SampleCode\MySqlTypeSamples\Edit
+     * @var \App\Service\Controller\Other\SampleCode\MySqlTypeSamples\Edit
      */
     private CtlService $ctlService;
 
@@ -74,7 +74,7 @@ class EditController extends AppController
             'input' => $this->ctlService->getInputProcess(),
         ]);
 
-        return $this->render('/SampleCode/MySqlTypeSamples/input');
+        return $this->render('/Other/SampleCode/MySqlTypeSamples/input');
     }
 
     /**
@@ -113,7 +113,7 @@ class EditController extends AppController
             'input' => $this->ctlService->getInputProcess(),
         ]);
 
-        return $this->render('/SampleCode/MySqlTypeSamples/conf');
+        return $this->render('/Other/SampleCode/MySqlTypeSamples/conf');
     }
 
     /**
