+++ src/Service/Controller/Other/SampleCode/MySqlTypeSamples/Shared/ValidatorSetting.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Service\Controller\SampleCode\MySqlTypeSamples\Shared;
+namespace App\Service\Controller\Other\SampleCode\MySqlTypeSamples\Shared;
 
 use App\Domain\Sample\MySqlTypeSamples\ValueObject as Vo;
 use App\Model\Table\Sample\MySqlTypeSamplesTable;
