+++ templates/Other/SampleCode/MySqlTypeSamples/detail.php
