+++ templates/Other/SampleCode/MySqlTypeSamples/input.php
